OptionalDeps: Split up conftest so that runtime-dependent utils live in tests/runtime/conftest
https://github.com/caikit/caikit/issues/228
