+++ tests/conftest.py
@@ -4,20 +4,15 @@ This sets up global test configs when pytest starts
 
 # Standard
 from contextlib import contextmanager
-from typing import Type
 from unittest.mock import patch
 import copy
 import json
 import os
 import sys
 import tempfile
-import threading
-import time
 import uuid
 
 # Third Party
-from grpc_health.v1 import health_pb2, health_pb2_grpc
-import grpc
 import pytest
 
 # First Party
@@ -25,14 +20,7 @@ import alog
 
 # Local
 from caikit import get_config
-from caikit.core.data_model.dataobject import render_dataobject_protos
 from caikit.core.toolkit import logging
-from caikit.runtime.grpc_server import RuntimeGRPCServer
-from caikit.runtime.model_management.model_manager import ModelManager
-from caikit.runtime.service_factory import ServicePackage, ServicePackageFactory
-from caikit.runtime.servicers.global_predict_servicer import GlobalPredictServicer
-from caikit.runtime.servicers.global_train_servicer import GlobalTrainServicer
-from tests.fixtures import Fixtures
 import caikit
 
 log = alog.use_channel("TEST-CONFTEST")
@@ -69,171 +57,19 @@ def test_environment():
     # No cleanup required...?
 
 
-@pytest.fixture(scope="session")
-def sample_inference_service(render_protos) -> ServicePackage:
-    """Service package pointing to `sample_lib` for testing"""
-    inference_service = ServicePackageFactory().get_service_package(
-        ServicePackageFactory.ServiceType.INFERENCE,
-    )
-    if render_protos:
-        render_dataobject_protos("tests/protos")
-    return inference_service
-
-
-@pytest.fixture(scope="session")
-def sample_predict_servicer(sample_inference_service) -> GlobalPredictServicer:
-    servicer = GlobalPredictServicer(inference_service=sample_inference_service)
-    yield servicer
-    # Make sure to not leave the rpc_meter hanging
-    # (It does try to clean itself up on destruction, but just to be sure)
-    servicer.rpc_meter.end_writer_thread()
-
-
-@pytest.fixture(scope="session")
-def sample_train_service(render_protos) -> ServicePackage:
-    """Service package pointing to `sample_lib` for testing"""
-    training_service = ServicePackageFactory().get_service_package(
-        ServicePackageFactory.ServiceType.TRAINING,
-    )
-    if render_protos:
-        render_dataobject_protos("tests/protos")
-    return training_service
-
-
-@pytest.fixture(scope="session")
-def sample_train_servicer(sample_train_service) -> GlobalTrainServicer:
-    servicer = GlobalTrainServicer(training_service=sample_train_service)
-    yield servicer
-
-
-@contextmanager
-def runtime_grpc_test_server(*args, **kwargs):
-    """Helper to wrap creation of RuntimeGRPCServer in temporary configurations"""
-    with tempfile.TemporaryDirectory() as workdir:
-        temp_log_dir = os.path.join(workdir, "metering_logs")
-        temp_save_dir = os.path.join(workdir, "training_output")
-        os.makedirs(temp_log_dir)
-        os.makedirs(temp_save_dir)
-        with temp_config(
-            {
-                "runtime": {
-                    "metering": {"log_dir": temp_log_dir},
-                    "training": {"output_dir": temp_save_dir},
-                }
-            },
-            "merge",
-        ):
-            with RuntimeGRPCServer(*args, **kwargs) as server:
-                # Give tests access to the workdir
-                server.workdir = workdir
-                yield server
-
-
-@pytest.fixture(scope="session")
-def runtime_grpc_server(
-    sample_inference_service, sample_train_service
-) -> RuntimeGRPCServer:
-    with runtime_grpc_test_server(
-        inference_service=sample_inference_service,
-        training_service=sample_train_service,
-    ) as server:
-        _check_server_readiness(server)
-        yield server
-
-
-@pytest.fixture(scope="session")
-def inference_stub(sample_inference_service, runtime_grpc_server) -> Type:
-    inference_stub = sample_inference_service.stub_class(
-        runtime_grpc_server.make_local_channel()
-    )
-    return inference_stub
-
-
-@pytest.fixture(scope="session")
-def train_stub(sample_train_service, runtime_grpc_server) -> Type:
-    train_stub = sample_train_service.stub_class(
-        runtime_grpc_server.make_local_channel()
-    )
-    return train_stub
-
-
-@pytest.fixture(scope="session")
-def training_management_stub(runtime_grpc_server) -> Type:
-    training_management_service: ServicePackage = (
-        ServicePackageFactory().get_service_package(
-            ServicePackageFactory.ServiceType.TRAINING_MANAGEMENT,
-        )
-    )
-
-    training_management_stub = training_management_service.stub_class(
-        runtime_grpc_server.make_local_channel()
-    )
-    return training_management_stub
-
-
 @pytest.fixture
 def good_model_path() -> str:
-    return Fixtures.get_good_model_path()
+    return os.path.join(FIXTURES_DIR, "models", "foo")
 
 
 @pytest.fixture
 def streaming_model_path() -> str:
-    return os.path.join(os.path.dirname(__file__), "fixtures", "dummy_streaming_module")
+    return os.path.join(FIXTURES_DIR, "dummy_streaming_module")
 
 
 @pytest.fixture
 def other_good_model_path() -> str:
-    return Fixtures.get_other_good_model_path()
-
-
-@pytest.fixture
-def sample_task_model_id(good_model_path) -> str:
-    """Loaded model ID using model manager load model implementation"""
-    model_id = _random_id()
-    model_manager = ModelManager.get_instance()
-    # model load test already tests with archive - just using a model path here
-    model_manager.load_model(
-        model_id,
-        local_model_path=good_model_path,
-        model_type=Fixtures.get_good_model_type(),  # eventually we'd like to be determining the type from the model itself...
-    )
-    yield model_id
-
-    # teardown
-    model_manager.unload_model(model_id)
-
-
-@pytest.fixture
-def streaming_task_model_id(streaming_model_path) -> str:
-    """Loaded model ID using model manager load model implementation"""
-    model_id = _random_id()
-    model_manager = ModelManager.get_instance()
-    model_manager.load_model(
-        model_id,
-        local_model_path=streaming_model_path,
-        model_type=Fixtures.get_good_model_type(),
-    )
-    yield model_id
-
-    # teardown
-    model_manager.unload_model(model_id)
-
-
-@pytest.fixture
-def other_task_model_id(other_good_model_path) -> str:
-    """Loaded model ID using model manager load model implementation"""
-    model_id = _random_id()
-    model_manager = ModelManager.get_instance()
-    # model load test already tests with archive - just using a model path here
-    model_manager.load_model(
-        model_id,
-        local_model_path=other_good_model_path,
-        model_type=Fixtures.get_good_model_type(),  # eventually we'd like to be determining the type from the model itself...
-    )
-    yield model_id
-
-    # teardown
-    model_manager.unload_model(model_id)
+    return os.path.join(FIXTURES_DIR, "models", "bar")
 
 
 @contextmanager
@@ -258,12 +94,15 @@ def temp_config(config_overrides: dict, merge_strategy="override"):
 
 # fixtures to optionally generate the protos for easier debugging
 def pytest_addoption(parser):
-    parser.addoption(
-        "--render-protos",
-        action="store_true",
-        default=False,
-        help="Render test protos for debug?",
-    )
+    try:
+        parser.addoption(
+            "--render-protos",
+            action="store_true",
+            default=False,
+            help="Render test protos for debug?",
+        )
+    except ValueError:
+        pass
 
 
 @pytest.fixture(scope="session")
@@ -312,7 +151,7 @@ def sample_int_file() -> str:
 
 @pytest.fixture
 def fixtures_dir():
-    yield os.path.join(os.path.dirname(os.path.realpath(__file__)), "fixtures")
+    yield FIXTURES_DIR
 
 
 @pytest.fixture
@@ -328,20 +167,3 @@ def toolkit_fixtures_dir(fixtures_dir):
 # IMPLEMENTATION DETAILS ############################################################
 def _random_id():
     return str(uuid.uuid4())
-
-
-def _check_server_readiness(server):
-    """Check server readiness"""
-
-    channel = grpc.insecure_channel(f"localhost:{server.port}")
-
-    done = False
-    while not done:
-        try:
-            stub = health_pb2_grpc.HealthStub(channel)
-            health_check_request = health_pb2.HealthCheckRequest()
-            stub.Check(health_check_request)
-            done = True
-        except grpc.RpcError:
-            log.debug("[RpcError]; will try to reconnect to test server in 0.1 second.")
-            time.sleep(0.1)
